+++ src/components/ExpenseItems.js
@@ -1,17 +1,10 @@
+import ExpenseDate from './ExpenseDate';
 import './ExpenseItems.css';
 
 function ExpenseItem(props) {
-	const month = props.date.toLocaleString('en-US', { month: 'long' });
-	const year = props.date.getFullYear();
-	const day = props.date.toLocaleString('en-US', { day: '2-digit' });
-
 	return (
 		<div className='expense-item'>
-			<div>
-				<div>{month}</div>
-				<div>{year}</div>
-				<div>{day}</div>
-			</div>
+			<ExpenseDate date={props.date} />
 
 			<div className='expense-item__description'>
 				<h2>{props.title}</h2>
